log workout

+++ app/src/main/java/com/example/healthcompass/data/FitnessActivity/FitnessActivity.kt
@@ -3,7 +3,7 @@ package com.example.healthcompass.data.FitnessActivity
 data class FitnessActivity(
     val activityName: String,
     val activityDate: String,
-    val caloriesBurnt: Double,
+    val caloriesBurnt: Int,
     val startTime: String,
     val endTime: String,
     val duration: String,
